reorganize common io functions
